requirements.txt
urllib3
requests
selenium
beautifulsoup4
tweepy
python-dotenv
translate